reqrequirements.txt
Django==1.10.1
django-cors-middleware==1.3.1
django-filter==0.15.0
djangorestframework==3.4.7
Markdown==2.6.6
wheel==0.26.0
